mongoDB/requirements.txt
Flask
pymongo
dnspython
numpy
faiss-cpu
requests